src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-reprojection-complex-expected.txt

CONSOLE ERROR: line 15: Uncaught TypeError: host1.createShadowRoot is not a function
